fastapi/requirements.txt
fastapi[standard]
uvicorn
databases[asyncpg]
pydantic
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
